alpha_signal_engine/requirements.txt
pandas
duckdb
pyyaml
lightgbm==3.3.5
matplotlib
jinja2
numpy
scipy
joblib
requests
# python-telegram-bot==13.15   ← DELETE or comment this
mlflow==2.17.2
sagemaker-mlflow==0.1.0
